WIP: Accessibility for Wiki modal dialogs
1) The wiki currently uses backbone.modal for its modal dialogs.
2) We already have accessibility code written for a standard format
leanModal dialog for use elsewhere in the site.

Proposal: Move Wiki modal dialogs to leanModel instead of backbone.

Problem 1: The preview button uses a button (rather than a link), and
then does some onclick jiggery-pokery to load the preview into an iframe
into the modal dialog.
Problem 2: The LMS uses a hacked up version of leanModal that has lots
of references to lms-specific things, and replaces the actual vender
leanModal code.

Proposal: Move back to vendor supplied leanModal in the LMS

Problem 3: We'll need to re-write the video popup specific stuff in the
current leanModal implementation.

[LMS_1303] [LMS-1337]